Revert "catch the com.webos.service.camera warnings"
This reverts commit a4862cf9f9f7e801f54c1fc498888bef6e8bac61.

+++ src/hal/v4l2/v4l2_camera_plugin.h
@@ -72,7 +72,7 @@ extern "C"
     void createFourCCPixelFormatMap();
     void createCameraPixelFormatMap();
     unsigned long getFourCCPixelFormat(camera_pixel_format_t);
-    camera_pixel_format_t getCameraPixelFormat(unsigned long);
+    camera_pixel_format_t getCameraPixelFormat(int);
 
     static int xioctl(int, int, void *);
